metodo delete resolvido foi um erro de build no banco de dados, como eu tinha colocado no carrinhoItem o produto como possivel de ser nulo o entityframework nao colocou o DELETE EM CASCADA na tabela, por isso eu não conseguia excluir itens que estavam associados, resolveu quando tirei o nulavel do produto

+++ AluraGeekAPI/Controllers/ProdutosController.cs
@@ -77,8 +77,15 @@ namespace AluraGeekAPI.Controllers
             {
                 return NotFound();
             }
+
+            if (produtoItem != null)
+            {
             _context.CarrinhoItens.Remove(produtoItem);
+               
+            }
+
             _context.Produtos.Remove(produto);
+
             await _context.SaveChangesAsync();
 
             return Ok("Item excluido");
